code for our first prototype with the chassis driving straight
motors were labelled to more accurately reflect the robot a whole, allowing the chassis drive to be fixed so it should drive straight. This change has yet to be tested, but should work.
the superstructure code is for our first prototype of a basic arcing lift (as found on k-9 from Starstruck)