Modifications for the new claw and addtion of new alternate auton and addtion of function to convert rotations to inches
made changes to the code to make to adapt the to the new gripping claw. Eleminated the function that counts rotations in degrees for the claw. added a new auton code that just drives back to flip the cap and then drives forwards to get out of the way. added a function to make calculations for the distance in inches easier on programmers